% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{Analyse}
\alias{Analyse}
\title{Fonction d'analyse}
\usage{
Analyse(heatmap, NbGrp, Data)
}
\arguments{
\item{heatmap}{Heatmap avec dendrograme générée par la fonction heatmap.}

\item{NbGrp}{le nombre de groupe que l'on souhaite analyser}

\item{Data}{le jeu de donnée néttoyé utilisé dans la fonction RF}
}
\value{
Les analyses
}
\description{
anlyse la Heatmap
}
