% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arbre.R
\name{Arbre}
\alias{Arbre}
\title{Fonction Arbre DIV}
\usage{
Arbre(Data, K, mtry)
}
\arguments{
\item{Data}{le jeu de donnée nétoyée ( QUANTI)}

\item{K}{le nombre de cluster attendu par Divclust.}

\item{mtry}{le nombre de variables prises en comptes}
}
\value{
Matrices de similarité, dissimilarité et absence.
}
\description{
Construit l'arbre divclust et l'analyse
}
