% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Fonction Clean}
\usage{
clean(Dat, centr = TRUE)
}
\arguments{
\item{Dat}{le jeu de donnée brut initial.}

\item{centr}{si on veux scale le jeu oui ou non.}
}
\value{
Le jeu de donné nétoyé : seulement les colonnes quanti et seulement les colonnes quali.
}
\description{
Prépare le dataset à l'utilisation de divclust avec RF
}
