% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heat}
\alias{heat}
\title{Foction Heatmap}
\usage{
heat(Forest)
}
\arguments{
\item{Forest}{le résultats de la fonction apply de sur la fonction RF ( arbre )}
}
\value{
Fourni les 2 type de heatmap + la matrice de similarité cumulée de la RF
}
\description{
Merge toutes les matrices et efectue la Heatmap sur la matrice de similarité.
}
