% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{mytree}
\alias{mytree}
\title{Fonction tree DIV}
\usage{
mytree(X, K, mtry)
}
\arguments{
\item{X}{le jeu de donnée nétoyée ( QUANTI)}

\item{K}{le nombre de cluster attendu par Divclust.}

\item{mtry}{le nombre de variables prises en comptes}
}
\value{
Matrices de similarité, dissimilarité et absentence.
}
\description{
Construit l'tree divclust et l'analyse
}
