% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rfclust.R
\name{plot.rfclust.summary}
\alias{plot.rfclust.summary}
\title{S3 plot method for rfclust.summary objects}
\usage{
\method{plot}{rfclust.summary}(x, ...)
}
\arguments{
\item{x}{the cumulative similarity matrix}

\item{...}{other arguments to be passed to and from other methods.}
}
\value{
Provides the associated heatmap of the similarity matrix
}
\description{
S3 plot method for rfclust.summary objects
}
