% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rfclust.R
\name{print.rfclust}
\alias{print.rfclust}
\title{S3 print method for rfclust objects}
\usage{
\method{print}{rfclust}(x, ...)
}
\arguments{
\item{x}{a \code{rfclust}.}

\item{...}{other arguments to be passed to and from other methods.}
}
\description{
S3 print method for rfclust objects
}
