% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.rfclust.R
\name{print.rfclust.summary}
\alias{print.rfclust.summary}
\title{S3 print method for rfclust.summary objects}
\usage{
\method{print}{rfclust.summary}(x, ...)
}
\arguments{
\item{x}{a \code{rfclust.summary}.}

\item{...}{other arguments to be passed to and from other methods.}
}
\description{
S3 print method for rfclust.summary objects
}
