% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfclust.R
\name{rfclust}
\alias{rfclust}
\title{Fonction générale}
\usage{
rfclust(
  X,
  n_trees = 500,
  K = 2,
  mtry = 1,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{X}{a data.frame to be clustered}

\item{n_trees}{number of \code{divlust} tress. Default is \code{500}.}

\item{K}{Nombre de cluster attendu par divclust.}

\item{mtry}{Nombre de variables prises en comptes.}
}
\value{
la heat map et son dendrograme.
}
\description{
Lie les fonctions ensemble j'espère mdr
}
