% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rfclust.R
\name{sum.rfclust}
\alias{sum.rfclust}
\title{Merge toutes les matrices}
\usage{
\method{sum}{rfclust}(x, ...)
}
\arguments{
\item{x}{le résultats de la fonction apply de sur la fonction RF ( tree )}

\item{...}{more parameters if required}
}
\value{
Fourni les 2 type de heatmap + la matrice de similarité cumulée de la RF
}
\description{
Merge toutes les matrices
}
