% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Fonction d'analyse}
\usage{
analyse(heatmap, nb_grp, X)
}
\arguments{
\item{heatmap}{Heatmap avec dendrograme générée par la fonction heatmap.}

\item{nb_grp}{le nombre de groupe que l'on souhaite analyser}

\item{X}{le jeu de donnée néttoyé utilisé dans la fonction RF}
}
\value{
Les analyses
}
\description{
anlyse la Heatmap
}
