% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Fonction Clean}
\usage{
clean(X)
}
\arguments{
\item{X}{Le jeu de donnée brut}
}
\value{
Le jeu de donné nétoyé : seulement les colonnes quanti et seulement les colonnes quali.
}
\description{
Prépare le dataset à l'utilisation de divclust avec RF
}
