% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfclust.R
\name{rfclust}
\alias{rfclust}
\title{Fonction générale}
\usage{
rfclust(X, center = TRUE, n_trees, K, mtry)
}
\arguments{
\item{X}{Le jeu de donnée brut}

\item{center}{Scale le jeu de donnée ?}

\item{n_trees}{Nombre d'trees pour la forêt.}

\item{K}{Nombre de cluster attendu par divclust.}

\item{mtry}{Nombre de variables prises en comptes.}
}
\value{
la heat map et son dendrograme.
}
\description{
Lie les fonctions ensemble j'espère mdr
}
