% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePack.R
\name{RFCLUST}
\alias{RFCLUST}
\title{Fonction générale}
\usage{
RFCLUST(Data_Raw, Center = TRUE, Nb_T, K, mtry)
}
\arguments{
\item{Data_Raw}{Le jeu de donnée brut}

\item{Center}{Scale le jeu de donnée ?}

\item{Nb_T}{Nombre d'arbres pour la forêt.}

\item{K}{Nombre de cluster attendu par divclust.}

\item{mtry}{Nombre de variables prises en comptes.}
}
\value{
la heat map et son dendrograme.
}
\description{
Lie les fonctions ensemble j'espère mdr
}
