% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rfclust.R
\name{summary.rfclust}
\alias{summary.rfclust}
\title{Merge all matrices and some analysis}
\usage{
\method{summary}{rfclust}(x, ...)
}
\arguments{
\item{x}{the results of the apply function on the RF function (tree)}

\item{...}{more parameters if required}
}
\value{
Provides the cumulative similarity matrix and some analysis
}
\description{
Merge all matrices and some analysis
}
